Refactor experiments loading into a context provider
